% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mervdown-package.R
\docType{package}
\name{mervdown-package}
\alias{mervdown}
\alias{mervdown-package}
\title{mervdown: R Markdown Templates by Merv}
\description{
A set of opinionated R Markdown templates for biological data analysis.
}
\examples{
library(rmarkdown)

available_templates("mervdown")

\dontrun{
## start an analysis with plyranges
draft("0_plyranges_analysis", "plyranges", "mervdown")

## start an analysis with SingleCellExperiment
draft("1_scRNAseq_analysis", "sce", "mervdown")

## start an analysis with tidyverse packages
draft("2_ggplots", "tidyverse", "mervdown")
}

}
\author{
\strong{Maintainer}: Mervin Fansler \email{fanslerm@mskcc.org} (\href{https://orcid.org/0000-0002-4108-4218}{ORCID})

}
