% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mervdown.R
\docType{package}
\name{mervdown}
\alias{mervdown}
\title{mervdown: R Markdown Templates by Merv}
\description{
A set of opinionated R Markdown templates for scientific data analysis.
}
\examples{
library(rmarkdown)

## start an analysis with plyranges
draft("0_plyranges_analysis", "plyranges", "mervdown")

## start an analysis with SingleCellExperiment
draft("1_scRNAseq_analysis", "sce", "mervdown")

## start an analysis with tidyverse packages
draft("2_ggplots", "tidyverse", "mervdown")

}
